test: 변경된 로직에 따라 테스트 케이스 수정
